add defens to DeleteCommentView view

+++ bwblog/views.py
@@ -78,6 +78,9 @@ class DeleteCommentView(DeleteView):
 
     def post(self, request, pk, slug, *args, **kwargs):
         comment = get_object_or_404(Comment, pk=pk)
-        comment.delete()
-
+        if comment.name == request.user.username:
+            comment.delete()
+            messages.add_message(request, messages.SUCCESS, 'Comment deleted!')
+        else:
+            messages.add_message(request, messages.ERROR, 'You can only delete your own comments!')
         return HttpResponseRedirect(reverse('post_detail', args=[slug]))
